updates in searchbar and cart page
