+++ pubspec.yaml
@@ -29,10 +29,10 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
   provider: ^4.3.3
-  share: 2.0.1
   google_sign_in: ^5.0.4
   firebase_auth: ^1.4.1
   firebase_core: ^1.3.0
+  share_plus: ^2.1.4
   
 
 dev_dependencies:
